SaidLine moves home, and the two moods come with it
Fourteen modules across eleven feature directories drew a line out of
`edit/`, and the `Said` type they mint came out of `edit/undoing.ts` —
undo's own module — because a row's editor and then ⌘Z were the first two
surfaces to need each. Neither is a feature's any more, which is exactly
the argument the same refactor pass used to hoist `refusals.tsx` to the
client's top level and did not act on here.

`client/SaidLine.tsx` is the component. `client/saying.ts` — already the
top-level receptacle for how long a said-line lingers, and already the
declared home of `SAID_MS` in everything but the file it was written in —
takes `Said` and the constant back beside it.

The two are deliberately not ONE file. Half this client mints a `Said`
without ever drawing one (a verb answers with a sentence and hands it to
whichever line is open), and a module with JSX in it cannot be imported
by a `bun test` unit test in this repo — the constraint `edit/undoing.ts`
documents about itself.

Behaviour-preserving: every importer's path, the two `claims.test.ts`
sweeps that name the files by path (`SAID_MS` collapses from two names to
one), and the prose that pointed at either.
