+++ packages/web/src/client/complete/Completions.tsx
@@ -42,7 +42,7 @@ import { Popper } from "@kobalte/core/popper"
 import { createSignal, Index, Show } from "solid-js"
 import { Portal } from "solid-js/web"
 
-import { ALARM_BAND, SaidLine } from "../edit/SaidLine.tsx"
+import { ALARM_BAND, SaidLine } from "../SaidLine.tsx"
 import { LAYER } from "../layer.ts"
 import { overlayRoot } from "../overlay.ts"
 import { Result, type RowTestids } from "../search/Result.tsx"
@@ -112,7 +112,7 @@ export function Completions(props: { readonly listing: Listing }) {
         {/* The search's own refusal, in its own words and in its own slot —
             never dropped, and never overwriting a list somebody is reading.
             The band is the one the other two shortlist panels wear and the
-            mood is `../edit/SaidLine.tsx`'s, once: this row spelled both by
+            mood is `../SaidLine.tsx`'s, once: this row spelled both by
             hand, down to the pad, and so carried no `data-tone` for a
             scenario to read the mood off. */}
         <Show when={props.listing.failure()}>
