+++ packages/web/src/client/date/RepeatPicker.tsx
@@ -31,7 +31,7 @@ import { createSignal, For } from "solid-js"
 
 import type { Press } from "../edit/panel.ts"
 import { RowPanel } from "../edit/RowPanel.tsx"
-import type { Said } from "../edit/undoing.ts"
+import type { Said } from "../saying.ts"
 import { TESTID } from "../testids.ts"
 import { TARGET } from "../touch.ts"
 import { noticeOf, pressOf, startsAt } from "./repeat.ts"
