+++ packages/web/src/client/menu/actions.ts
@@ -38,7 +38,8 @@ import type { Shelf } from "@olai/surface"
 
 import { armNode } from "../chat/armed.ts"
 import type { Relation } from "../edges/relation.ts"
-import type { Said, Undo } from "../edit/undoing.ts"
+import type { Said } from "../saying.ts"
+import type { Undo } from "../edit/undoing.ts"
 import { setFolded } from "../fold/memory.ts"
 import { type Fold, foldIdOf, foldOf } from "../fold/rows.ts"
 import { setChatOpen } from "../layout/prefs.ts"
