+++ packages/web/src/client/props/PropEditor.tsx
@@ -22,7 +22,7 @@ import { createSignal } from "solid-js"
 
 import type { Press } from "../edit/panel.ts"
 import { RowPanel } from "../edit/RowPanel.tsx"
-import type { Said } from "../edit/undoing.ts"
+import type { Said } from "../saying.ts"
 import { TESTID } from "../testids.ts"
 import { TARGET } from "../touch.ts"
 import { type Editing, pressOf } from "./editor.ts"
