+++ packages/web/src/client/search/Shortlist.tsx
@@ -53,7 +53,7 @@
 import type { NodeHit } from "@olai/surface"
 import { type Accessor, createMemo, createSignal, Index, onMount, Show } from "solid-js"
 
-import { SaidLine } from "../edit/SaidLine.tsx"
+import { SaidLine } from "../SaidLine.tsx"
 import { listKey } from "../keys.ts"
 import { Refused } from "../Refused.tsx"
 import type { TestId } from "../testids.ts"
@@ -290,7 +290,7 @@ export function Shortlist(props: {
           the same slot as a hit's: one is the server saying it could not
           answer, the other is an answer. A LINE under the box rather than a
           band across a panel, which is the filter bar's shape and not the two
-          popovers'; the mood is `../edit/SaidLine.tsx`'s either way. */}
+          popovers'; the mood is `../SaidLine.tsx`'s either way. */}
       <Show when={found.failure()}>
         {(failure) => (
           <SaidLine
